test(groovebox): move shared registry fixtures to tests/helpers (importing a test file re-ran its describes)
