+++ docs/arcade/groovebox/tests/registry-validate.test.js
@@ -10,21 +10,7 @@ import { digitalLove } from '../songs/digital-love.js';
 import { memoryReboot } from '../songs/memory-reboot.js';
 import { takeOnMe } from '../songs/take-on-me.js';
 
-// Minimal valid payloads — also imported by registry-handler.test.js
-export const GROOVE_PAYLOAD = {
-  laneType: 'drums',
-  meter: { beatsPerBar: 4, beatUnit: 4, stepsPerBeat: 4 },
-  bars: [{ kick: [0, 8], snare: [4, 12], hat: [0, 2, 4, 6, 8, 10, 12, 14], crash: [], tom: [] }],
-};
-export const SONG_PAYLOAD = {
-  version: 2, title: 'T', artist: 'A',
-  meter: { beatsPerBar: 4, beatUnit: 4, stepsPerBeat: 4 }, bpm: 120,
-  lanes: [{ id: 'drums', type: 'drums', name: 'Drums', muted: false, soloed: false }],
-  grooves: { drums: { main: [{ kick: [0], snare: [], hat: [], crash: [], tom: [] }] } },
-  patterns: [{ lanes: { drums: 'main' } }],
-  chain: [0],
-  fills: {},
-};
+import { GROOVE_PAYLOAD, SONG_PAYLOAD } from './helpers/registry-fixtures.js';
 const CREATE = (over = {}) => ({
   kind: 'groove', schema_version: 1, name: 'amen-ish', author: 'Henry',
   payload: GROOVE_PAYLOAD, ...over,
